+++ back/src/bsds/resolvers/queries/__tests__/bsds.bsvhu.integration.ts
@@ -21,7 +21,11 @@ import {
 } from "../../../../../integration-tests/helper";
 import makeClient from "../../../../__tests__/testClient";
 import { ErrorCode } from "../../../../common/errors";
-import { indexBsvhu } from "../../../../bsvhu/elastic";
+import {
+  BsvhuForElasticInclude,
+  getBsvhuForElastic,
+  indexBsvhu
+} from "../../../../bsvhu/elastic";
 import {
   transporterReceiptFactory,
   userWithCompanyFactory
@@ -536,9 +540,14 @@ describe("Query.bsds.vhus base workflow", () => {
     beforeAll(async () => {
       const refusedVhu = await prisma.bsvhu.update({
         where: { id: vhuId },
-        data: { status: "REFUSED" }
+        data: { status: "REFUSED" },
+        include: {
+          ...BsvhuForElasticInclude
+        }
       });
-      await indexBsvhu(refusedVhu);
+      const bsvhuForElastic = await getBsvhuForElastic(refusedVhu);
+      await indexBsvhu(bsvhuForElastic);
+
       await refreshElasticSearch();
     });
 
@@ -614,7 +623,8 @@ describe("Query.bsds.vhus mutations", () => {
       }
     });
 
-    await indexBsvhu(vhu);
+    const bsvhuForElastic = await getBsvhuForElastic(vhu);
+    await indexBsvhu(bsvhuForElastic);
     await refreshElasticSearch();
 
     const { query } = makeClient(emitter.user);
@@ -663,7 +673,8 @@ describe("Query.bsds.vhus mutations", () => {
         emitterCompanySiret: emitter.company.siret
       }
     });
-    await indexBsvhu(vhu);
+    const bsvhuForElastic = await getBsvhuForElastic(vhu);
+    await indexBsvhu(bsvhuForElastic);
 
     //duplicate vhu
     const { mutate } = makeClient(emitter.user);
@@ -689,7 +700,7 @@ describe("Query.bsds.vhus mutations", () => {
       }
     });
 
-    await indexBsvhu(vhu);
+    await indexBsvhu(bsvhuForElastic);
     await refreshElasticSearch();
 
     const { query } = makeClient(emitter.user);