+++ back/src/registry/__tests__/where.dates.integration.ts
@@ -9,7 +9,7 @@ import { getBsdasriForElastic, indexBsdasri } from "../../bsdasris/elastic";
 import { bsdasriFactory } from "../../bsdasris/__tests__/factories";
 import { getBsffForElastic, indexBsff } from "../../bsffs/elastic";
 import { createBsff } from "../../bsffs/__tests__/factories";
-import { indexBsvhu } from "../../bsvhu/elastic";
+import { getBsvhuForElastic, indexBsvhu } from "../../bsvhu/elastic";
 import { bsvhuFactory } from "../../bsvhu/__tests__/factories.vhu";
 import { client, index } from "../../common/elastic";
 import { getFormForElastic, indexForm } from "../../forms/elastic";
@@ -267,8 +267,8 @@ describe("toElasticFilter", () => {
     });
 
     await Promise.all(
-      [bsvhu1, bsvhu2, bsvhu3, bsvhu4].map(async bsda => {
-        return indexBsvhu(bsda);
+      [bsvhu1, bsvhu2, bsvhu3, bsvhu4].map(async bsvhu => {
+        return indexBsvhu(await getBsvhuForElastic(bsvhu));
       })
     );
     await refreshElasticSearch();
@@ -703,8 +703,8 @@ describe("toElasticFilter", () => {
     });
 
     await Promise.all(
-      [bsvhu1, bsvhu2, bsvhu3, bsvhu4].map(async bsda => {
-        return indexBsvhu(bsda);
+      [bsvhu1, bsvhu2, bsvhu3, bsvhu4].map(async bsvhu => {
+        return indexBsvhu(await getBsvhuForElastic(bsvhu));
       })
     );
     await refreshElasticSearch();
@@ -1186,8 +1186,8 @@ describe("toElasticFilter", () => {
     });
 
     await Promise.all(
-      [bsvhu1, bsvhu2, bsvhu3, bsvhu4].map(async bsda => {
-        return indexBsvhu(bsda);
+      [bsvhu1, bsvhu2, bsvhu3, bsvhu4].map(async bsvhu => {
+        return indexBsvhu(await getBsvhuForElastic(bsvhu));
       })
     );
     await refreshElasticSearch();